backend/requirements/prod.txt
Django==5.1.2
psycopg2-binary==2.9.9
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Pillow==10.4.0
numpy==2.1.2

django-cors-headers==4.4.0
